[aarch64] Fixes CLI crashes on dpaa2 platform.
- always use 'va_args' as pointer in all format_* functions
- u32 for all 'indent' params as it's declaration was inconsistent

+++ src/vnet/adj/adj_glean.c
@@ -228,7 +228,7 @@ format_adj_glean (u8* s, va_list *ap)
     ip_adjacency_t * adj = adj_get(index);
 
     return (format(s, "%U-glean: %U",
-		   format_fib_protocol, adj->ia_nh_proto,
+		   format_fib_protocol, (int)adj->ia_nh_proto,
                    format_vnet_sw_interface_name,
                    vnm,
                    vnet_get_sw_interface(vnm,
