+++ src/vnet/fib/ip6_fib.c
@@ -601,7 +601,7 @@ ip6_show_fib (vlib_main_t * vm,
 
 	s = format(s, "%U, fib_index:%d, flow hash:[%U] locks:[",
                    format_fib_table_name, fib->index,
-                   FIB_PROTOCOL_IP6,
+                   (int)FIB_PROTOCOL_IP6,
                    fib->index,
                    format_ip_flow_hash_config,
                    fib_table->ft_flow_hash_config);
